New Fade Script and other changes.
Added a new script, Fade, that can be added to objects that have sprite or image renderers. Has a function that can be called to pass parameters and start the appropriate coroutine. It will then fade the sprite to the desired transparency over the chosen duration.  I then made changes to the player movement script and my test scene to make the virtual joystick fade out (mostly) when not in use.