Enemy JSON File
Created a scene to create enemies. Fill in information about enemy to add into game, such as ID, Name, Health, Damage, and then click "Add Enemy to Json" and if the json file does not exist, it will create the file, otherwise it will append to file.

(Scene is just for ease of adding to json file, so we do not have to add manually)