Lots of things
Fixed the virtual joystick in PlayerMovement. Created a folder for the random encounters. The random encounter object which I will push next will just be a sprite of tall grass or some other conspicuous object with a collider in front of it that will send you to the combat scene. Also added a few silly sprites and added the Encounter script to handle the scene transition. Also made one tiny change to the PauseMenu script so that isPaused will be set to false when play is resumed.