Encounters
There is now an encounter object that can be placed in a level to send the player to the combat scene. Need to iron out the details of the transition. Also created a new folder for scripts and made some changes to my test scene. The encounter object will disappear once activated and the player moves a certain distance away from it, but you are currently unable to   return from combat to the main game world.