PlayerStatus script for encounters
Added a PlayerStatus script that can be used to track the world position of the player and their status. This script has two functions that can be called to transition the player from the world scene to combat and back again respectively. There are two similar functions for simply entering a new area that would be in a different scene (like in a cave or building).