Cross-Scene Player Object and Preload Scene
Created a preload scene and accompanying script that contains the player and (temporarily) the RandomEncounter and will pass them on to my test scene using DontDestroyOnLoad. Also created a VirtualJoystick script to separate the functionality of the virtual joystick out of the player scripts and into the joystick itself (which fixed issues with having the player start in the preload scene). Also added functionality to the flee option in combat so it will return to my test scene. Also added the preload scene to those in the build settings, so when this inevitably causes issues just disable it in the build settings and we can figure it out later.