Scene transitions and audio!
Added a black panel and SceneTransition script that utilizes the fade script to fade a black panel to create scene transitions. Also created a UniversalAudio object and a corresponding script to handle music transitions. Also got rid of the enemy prefabs and added the generic enemy capsule as a template and updated the JoelTestScene accordingly to use it. Updated various folders in relation to these changes as well.