Created Dev menu
Created a dev menu (all references in game refer to it as debug menu). The menu can be used to: disable player collision, set the timescale to 10X the normal time scale, and enter God-mode (invincibility and insta-kills in combat). Can open and close the menu  using the '=' key. The debug menu itself is created in the preload scene and will persist between scenes.