Debug search bar
Added a search bar to the debug menu that can be used to navigate to any scene while in game. Does not include error handling, so it is up to you guys to make sure you type the scene in correctly and type in scenes that are in the build order.