Audio Rework
Reworked the way that universal audio worked. It no longer exists and now is just a script on the world player. Also added a simple footstep system that can be expanded later. Added sound to the UI (partially functional). Also made a few random changes to fix crashes that were happening.