Disabled player movement in dialogue
Made changes to the dialogue scripts as well as the player status and player movement scripts to disable the player from moving during dialogue. Also implemented this change for when the player opens the quest menu. Also fixed some minor bugs in the aforementioned scripts. Also updated the player prefab to have the animation components given to the player that was being used for testing in the game world scene. I then removed the test player from the game world scene and made other minor adjustments so the scene in Github will be in the intended state.