Jeff moves now
Revamped the way the NPCMovement script worked to be more robust. Also created a Jeff movement script that specifies when and how Jeff should move and calls methods from NPCMovement to make Jeff move. Also created a public method in the Dialogue manager so I could grab an NPC's value and updated some comments in files I've worked on.