Added Prefab for scene transition
There is a new prefab called the TransitionPoint. It can be placed anywhere you would like the player to be able to go to a different scene. Simply enter the build index of the scene you would like it to take the player and you are good. You can update the message in its canvas to tell the player where they are heading, and you can move the entry and exit points to determine where the player will be when they enter from that scene and where they will move towards as they exit the current scene respectively. It also handles the fadepanel and all other necessary scene transtition (except for music change). Speaking of the fade panel, you can move it to wherever you want because it will now automatically position itself at the start of the scene, so it no longer needs disabled. Made a few other minor changes. The player now keeps track of the previous scene they were in (except for the combat scene).