Medley of changes
Removed the SetScenePosition script. Added the thyme road scene to the build order. Updated the time for scene transitions to 2 seconds (both the fade effect and the actual time before the change occurs). Fixed a minor error with scene transitions. Also got rid of the death button in the thyme road scene. Also added transition points in startville and thyme road so that the player can navigate between the two scenes.