Fixed gameplay scene not starting
Fixed two issues preventing me from starting the gameplay scene. 1) I updated the build indexes in the audio scripts to reflect the proper build indexes for the gameplay scene (and made a minor change to one of the scripts). 2) Restored the canvas of the world transition point that was added in front of the church. That child object gets referenced by the script and it will throw an error if you delete it. Do not delete stuff from prefabs randomly. If you wish for there to be no message, set the text to nothing and disable the button.