Enemy audio in combat
Other than the final boss and the dungeon boss, every enemy has audio for every action they can take (not necessarily different audio). There is an issue where crit attacks will play the sound corresponding to the reg attack.  The issue seems to be down to how I coded the audio and design decisions related to the classification of enemy actions. This will require a small rework of existing systems so it is going on the backburner until more pressing issues are resolved. Also moved all of the sound effects to the resources folder