Beginnings of Help UI in Combat
Began working on a new ui for giving the player info in combat. Also made the balance changes to inspire and healing winds I suggested. Made minor adjustments to the position of some ui elements in the combat scene also. Also made a Help script to be used to give functionality to this new ui.