Differentiated player and enemy shield stat
Enemies now have a new stat called protected that is represented by a helmet icon. The description of this stat describes the 50% damage reduction to incoming damage that the enemy receives. The player still has the shield stat represented by the shield icon which nullifies incoming damage with a 20% chance to fail.