Changes to player attacks
The time skip now is hardcoded to deal 5 damage, so it won't be affected by upgrades anymore. Also reworked the shield to only go down on a failed block. In other words, if your shield is successful in blocking the attack it is still active, but if it fails then it breaks and you have to wait for it to go off cooldown. I changed the failure chance of the shield from a 1/5 to a 1/4 to compensate, but this may need to be a bit more severe given the buff. Consequently increased the cooldown to 8 turns (6 might be sufficient).  Also decreased the damage multiplier of empower from 2.0 (double damage) to 1.5 (50% more damage). Most fights could be ended in two turns with empower + heavy axe.